Improved chain:summary-status
List of directories to report is now read from
$chain:summary-repos-file (default
~/.elvish/package-data/elvish-themes/chain-summary-repos.json). This
list can be manipulated using the new chain:add-summary-repo and
chain:remove-summary-repo functions.

The command used to find all git repos when the &all option is given
to chain:summary-status can now be customized by assigning a lambda to
$chain:find-all-user-repos.

chain:summary-status now uses the new `order` built-in introduced in
commit 1cd192daaf40ef46fd5378881ec832fa7faaaa85 to sort the list by
reverse updae timestamp. If the `order` function does not exist, the
list is presented unordered.

Internal change: the -parse-git function now supports a
&with-timestamp option (default $false) which, if true, calls the
corresponding command and adds a "timestamp" field to the last-status
data structure.