New $chain:show-last-chain option
By default, all segments are separated by the `chain` glyph. As a
special case, you can specify whether the last segment should be
separated by a chain or not by setting the `$chain:show-last-chain`
variable to `$false`. This can be used to designate some special
character in the =arrow= segment glyph (e.g. a Unicode arrow) and have
it shown without a separator. For example:

    chain:show-last-chain = $false
    chain:glyph[arrow] = "→"